% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2JSON.R
\name{convert2JSON}
\alias{convert2JSON}
\title{Returns a JSON-formatted string}
\usage{
convert2JSON(x, choice = "single")
}
\arguments{
\item{x}{The input list of the \code{\link{multiQuestions}} function, which is
a list of lists, each containing 3 or 4 elements (question, options,
answer, and optionally tip).}

\item{choice}{A character string specifying the quiz mode. Either \code{"single"}
(default) or \code{"multiple"} for multiple-choice.}
}
\value{
The function returns a JSON-formatted string containing a JS object of all
questions.
}
\description{
\code{convert2JSON} is a helper function for the \code{\link{multiQuestions}}
quiz function. It converts its input list into a Javascript object and returns
it as JSON text. This text can then be read by the multiQuestions Javascript code.
}
\seealso{
\code{\link{multiQuestions}}
}
\keyword{internal}
