% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillBlanks.R
\name{fillBlanks-shiny}
\alias{fillBlanks-shiny}
\alias{fillBlanksOutput}
\alias{renderFillBlanks}
\title{Shiny bindings for fillBlanks}
\usage{
fillBlanksOutput(outputId, width = "100\%", height = "500px")

renderFillBlanks(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a fillBlanks quiz}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using fillBlanks within Shiny
applications and interactive Rmd documents.
}
\examples{
\dontrun{
library(shiny)

txt <- list(
  cloze = "R is a $$!programming!$$ language for $$!statistical computing!$$.",
  addOptions = c("natural", "cooking")
)

ui <- fluidPage(
  fillBlanksOutput("quiz1")
)

server <- function(input, output) {
  output$quiz1 <- renderFillBlanks({
    fillBlanks(x = txt, title = "Fill in the Blanks")
  })
}

shinyApp(ui, server)
}

}
