% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiQuestions.R
\name{multiQuestions-shiny}
\alias{multiQuestions-shiny}
\alias{multiQuestionsOutput}
\alias{renderMultiQuestions}
\title{Shiny bindings for multiQuestions}
\usage{
multiQuestionsOutput(outputId, width = "100\%", height = "400px")

renderMultiQuestions(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a multiQuestions quiz}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
\code{multiQuestionsOutput} returns an HTML output element for use
  in a Shiny UI definition. \code{renderMultiQuestions} returns a
  server-side rendering function to be assigned to an output slot.
}
\description{
Output and render functions for using multiQuestions within Shiny
applications and interactive Rmd documents.
}
\examples{
if (interactive()) {
library(shiny)

x <- list(
  q1 = list(
    question = "What is 2+2?",
    options = c("3", "4", "5"),
    answer = 2
  ),
  q2 = list(
    question = "What is 3*3?",
    options = c("6", "9", "12"),
    answer = 2
  )
)

ui <- fluidPage(
  multiQuestionsOutput("quiz1")
)

server <- function(input, output) {
  output$quiz1 <- renderMultiQuestions({
    multiQuestions(x = x, title = "Math Quiz")
  })
}

shinyApp(ui, server)
}

}
