% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleQuestion.R
\name{singleQuestion-shiny}
\alias{singleQuestion-shiny}
\alias{singleQuestionOutput}
\alias{renderSingleQuestion}
\title{Shiny bindings for singleQuestion}
\usage{
singleQuestionOutput(outputId, width = "100\%", height = "400px")

renderSingleQuestion(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a singleQuestion.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using singleQuestion within Shiny
applications and interactive Rmd documents.
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  singleQuestionOutput("quiz1")
)

server <- function(input, output) {
  output$quiz1 <- renderSingleQuestion({
    singleQuestion(
      x = list("What is 2+2?", c("3", "4", "5"), 2),
      title = "Quick Quiz"
    )
  })
}

shinyApp(ui, server)
}

}
